Refact and added docs to sticker method
